# docs/swagger.yaml
definitions:
  entity.ConfigSpiderData:
    properties:
      cmd:
        description: 自定义爬虫
        type: string
      col:
        type: string
      display_name:
        type: string
      engine:
        description: 可配置爬虫
        type: string
      name:
        description: 通用
        type: string
      remark:
        type: string
      settings:
        type: object
      stages:
        items:
          $ref: '#/definitions/entity.Stage'
        type: array
      start_stage:
        type: string
      start_url:
        type: string
      type:
        type: string
    type: object
  entity.Field:
    properties:
      attr:
        type: string
      css:
        type: string
      name:
        type: string
      next_stage:
        type: string
      remark:
        type: string
      xpath:
        type: string
    type: object
...
  entity.Stage:
    properties:
      fields:
        items:
          $ref: '#/definitions/entity.Field'
        type: array
      is_list:
        type: boolean
      list_css:
        type: string
      list_xpath:
        type: string
      name:
        type: string
      page_attr:
        type: string
      page_css:
        type: string
      page_xpath:
        type: string
    type: object
  model.Env:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
...
  model.Spider:
    properties:
      _id:
        description: 爬虫ID
        type: string
      cmd:
        description: 自定义爬虫
        type: string
      col:
        description: 结果储存位置
        type: string
      config:
        $ref: '#/definitions/entity.ConfigSpiderData'
        description: 可配置爬虫配置
        type: object
      create_ts:
        type: string
      dedup_field:
        description: 去重字段
        type: string
      dedup_method:
        description: 去重方式
        type: string
      display_name:
        description: 爬虫显示名称
        type: string
      envs:
        description: 环境变量
        items:
          $ref: '#/definitions/model.Env'
        type: array
      file_id:
        description: GridFS文件ID
        type: string
      git_auto_sync:
        description: Git 是否自动同步
        type: boolean
      git_branch:
        description: Git 分支
        type: string
      git_has_credential:
        description: Git 是否加密
        type: boolean
      git_password:
        description: Git 密码
        type: string
      git_sync_error:
        description: Git 同步错误
        type: string
      git_sync_frequency:
        description: Git 同步频率
        type: string
      git_url:
        description: Git URL
        type: string
      git_username:
        description: Git 用户名
        type: string
      is_dedup:
        description: 去重
        type: boolean
      is_git:
        description: Git 设置
        type: boolean
      is_long_task:
        description: 长任务
        type: boolean
      is_public:
        description: 是否公开
        type: boolean
      is_scrapy:
        description: Scrapy 爬虫（属于自定义爬虫）
        type: boolean
      is_web_hook:
        description: Web Hook
        type: boolean
      last_run_ts:
        description: 前端展示
        type: string
      last_status:
        description: 最后执行状态
        type: string
      latest_tasks:
        description: 最近任务列表
        items:
          $ref: '#/definitions/model.Task'
        type: array
      name:
        description: 爬虫名称（唯一）
        type: string
      project_id:
        description: 项目ID
        type: string
      remark:
        description: 备注
        type: string
      site:
        description: 爬虫网站
        type: string
      spider_names:
        description: 爬虫名称列表
        items:
          type: string
        type: array
      src:
        description: 源码位置
        type: string
      template:
        description: 可配置爬虫
        type: string
      type:
        description: 爬虫类别
        type: string
      update_ts:
        type: string
      user_id:
        description: 时间
        type: string
      username:
        description: 用户名称
        type: string
      web_hook_url:
        description: Web Hook URL
        type: string
    type: object
  model.Task:
    properties:
      _id:
        type: string
      cmd:
        type: string
      create_ts:
        type: string
      error:
        type: string
      error_log_count:
        type: integer
      finish_ts:
        type: string
      log_path:
        type: string
      node_id:
        type: string
      node_name:
        type: string
      param:
        type: string
      pid:
        type: integer
      result_count:
        type: integer
      run_type:
        type: string
      runtime_duration:
        type: number
      schedule_id:
        type: string
      spider_id:
        type: string
      spider_name:
        description: 前端数据
        type: string
      start_ts:
        type: string
      status:
        type: string
      total_duration:
        type: number
      update_ts:
        type: string
      user_id:
        type: string
      username:
        type: string
      wait_duration:
        type: number
    type: object
...
info:
  contact: {}
  license: {}
paths:
  /config_spiders:
    put:
      consumes:
        - application/json
      description: Put config spider
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: spider item
          in: body
          name: spider
          required: true
          schema:
            $ref: '#/definitions/model.Spider'
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Put config spider
      tags:
        - config spider
  /config_spiders/{id}/config:
    get:
      consumes:
        - application/json
      description: Get config spider
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: spider id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Get config spider
      tags:
        - config spider
    post:
      consumes:
        - application/json
      description: Post config spider config
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: spider item
          in: body
          name: spider
          required: true
          schema:
            $ref: '#/definitions/model.Spider'
            type: object
        - description: spider id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Post config spider config
      tags:
        - config spider
  /config_spiders/{id}/spiderfile:
    post:
      consumes:
        - application/json
      description: Post config spider
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: spider id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Post config spider
      tags:
        - config spider
  /config_spiders/{id}/upload:
    post:
      consumes:
        - application/json
      description: Upload config spider
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: spider item
          in: body
          name: spider
          required: true
          schema:
            $ref: '#/definitions/model.Spider'
            type: object
        - description: spider id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Upload config spider
      tags:
        - config spider
  /config_spiders_templates:
    get:
      consumes:
        - application/json
      description: Get config spider template list
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: spider id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Get config spider template list
      tags:
        - config spider
...
  /nodes:
    get:
      description: Get nodes
      parameters:
        - description: With the bearer started
          in: header
          name: Authorization
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get nodes
      tags:
        - node
  /nodes/{id}:
    delete:
      description: Delete node
      parameters:
        - description: With the bearer started
          in: header
          name: Authorization
          required: true
          type: string
        - description: node id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Delete node
      tags:
        - node
    get:
      description: Get node
      parameters:
        - description: With the bearer started
          in: header
          name: Authorization
          required: true
          type: string
        - description: id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get node
      tags:
        - node
    post:
      consumes:
        - application/json
      description: Post node
      parameters:
        - description: With the bearer started
          in: header
          name: Authorization
          required: true
          type: string
        - description: post node
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Post node
      tags:
        - node
...
  /nodes/{id}/system:
    get:
      description: Get system info
      parameters:
        - description: With the bearer started
          in: header
          name: Authorization
          required: true
          type: string
        - description: node id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get system info
      tags:
        - node
  /nodes/{id}/tasks:
    get:
      description: Get tasks on node
      parameters:
        - description: With the bearer started
          in: header
          name: Authorization
          required: true
          type: string
        - description: node id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get tasks on node
      tags:
        - node
...
swagger: "2.0"